PFS/src/domainobjecttests/ExpenseTest.java
package domainobjecttests;

import java.util.Date;

import domainobjects.Expense;
import domainobjects.IDSet;
import domainobjects.PaymentMethod;
import junit.framework.TestCase;

public class ExpenseTest extends TestCase
{
	public void testCreation()
	{
		int[] setData = { 0 };
		Expense expense = new Expense(new Date(), 100, PaymentMethod.CASH, "Something to eat", 0, IDSet.createFromArray(setData));

		assertNotNull("Constructor for Expense failed", expense);
	}
	
	public void testCentsAndDollars()
	{
		int[] setData = { 0 };
		Expense expense = new Expense(new Date(), 105, PaymentMethod.CASH, "Something to eat", 0, IDSet.createFromArray(setData));
	
		assertEquals("Dollars are wrong", expense.getDollars(), 1);
		assertEquals("Cents are wrong", expense.getCents(), 5);
	}
	
	public void testGetters()
	{
		Date date = new Date();
		int cents = 105;
		PaymentMethod method = PaymentMethod.CASH;
		String description = "Something to eat";
		int payToID = 55;		
		
		int[] setData = { 1, 2, 3 };
		IDSet set = IDSet.createFromArray(setData);
		
		Expense expense = new Expense(date, cents, method, description, payToID, set);
		
		assertEquals("Date is not what was put in", expense.getDate(), date);
		assertEquals("Amount in cents is not what was put in", expense.getTotalAmountInCents(), cents);
		assertEquals("Payment method is not what was put in", expense.getPaymentMethod(), method);
		assertEquals("Description is not what was put in", expense.GetDescription(), description);
		assertEquals("Pay To is not what was put in", expense.getPayTo(), payToID);
		assertEquals("ID set is not what was put in", expense.getLabels(), set);
	}
}
